+++ src/components/MapFilters.vue
@@ -244,6 +244,9 @@
               >Show all active flooding</v-btn
             >
           </div>
+          <div style="display: none" id="noActiveFlooding">
+            No Active Flooding
+          </div>
           <input
             type="checkbox"
             ref="allRP"
@@ -601,7 +604,9 @@ export default {
 }
 
 #activeLayerTitle {
+  display: none;
   font-size: 14px;
+  padding-top: 10px;
 }
 
 #supporting-layer-title {
@@ -644,8 +649,14 @@ export default {
 }
 
 #showAllBtn {
-  display: flex;
+  display: none;
   justify-content: center;
   padding: 10px;
+}
+
+#noActiveFlooding {
+  text-align: left;
+  padding-top: 10px;
+  font-size: 14px;
 }
 </style>